requirements.txt
python-dotenv
python-telegram-bot[webhooks]==22.3
redis
requests
aiohttp